+++ ZeroApp.ForecastTracker.Service.Infrastructure/Entities/Location.cs
@@ -6,8 +6,8 @@
 
         public string Name { get; set; }
 
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
